refactor: 드로잉코어 이벤트 처리 분리
